refactor: 서비스/컨트롤러 새 응답 구조 적용 #5
- 새로운 ResponseCode 사용
- Dto 파일 위치 이동
