+++ src/main/java/com/second/kirby/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.second.kirby.dto;
+package com.second.kirby.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
