+++ src/main/java/com/second/kirby/dto/request/SignupRequest.java
@@ -1,4 +1,4 @@
-package com.second.kirby.dto;
+package com.second.kirby.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
