+++ src/main/java/com/second/kirby/exception/BusinessException.java
@@ -5,15 +5,15 @@ import lombok.Getter;
 @Getter
 public class BusinessException extends RuntimeException {
 
-    private final ErrorCode errorCode;
+    private final ResponseCode responseCode;
 
-    public BusinessException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
+    public BusinessException(ResponseCode responseCode) {
+        super(responseCode.getMessage());
+        this.responseCode = responseCode;
     }
 
-    public BusinessException(ErrorCode errorCode, String message) {
-        super(message);
-        this.errorCode = errorCode;
+    public BusinessException(ResponseCode responseCode, String customMessage) {
+        super(customMessage);
+        this.responseCode = responseCode;
     }
 }
